spotify_scraper/requirements.txt
playwright==1.39.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pandas>=2.1.1
pydantic==2.5.0
pytest==7.4.3
beautifulsoup4==4.12.2